разделил файл конфигурации н два(каждый отвечает за свою бд), добавил скрипт для добавления рекомендаций
